Implementing Unit Selection
Implemented click-and-drag unit selection in Gameplay, using a physics collision query; Currently doesn't draw a box to visualize what will be selected yet.
Implemented pressing Escape to deselect the selected units.
Implemented pressing F10 to bring up the main menu scene; Stores the current Gameplay scene so it can be resumed instead of restarted.

Added ControllableUnit node to represent a unit the player can select and, eventually, control. Selected status is based on the visibility of the SelectedOutline Sprite2D node; Currently using placeholder sprites for the main sprite and the outline sprite.

Added a ControllableUnit instance to Gameplay scene for testing.