Updated MainMenu
Added a Resume button to the MainMenu; It is only visible if the SceneManager currently has a stored scene.
Updated logic so that Start Game will free any stored scene before changing to the Gameplay scene.
Implemented Resume to restore the stored scene rather than changing to a new scene.