Implemented Selection Box
Implemented SelectionBox node which uses its _Draw method to draw a rectangle when it is Active.
Added SelectionBox node to Gameplay scene.
Updated Gameplay logic to update the SelectionBox's status on InputEventMouseMotion events; Only updates the rectangle while dragging, since not dragging == not active and so nothing is drawn.

Added a second ControllableUnit to Gameplay scene to test selecting and moving multiple units.