Implemented Simple Gameplay Loop
Added Level1 and Level2 scenes, using a TileMap to draw out a map to navigate. The simple gameplay loop is to simply move the unit into the marked goal. LevelBase provides a signal, LevelCompleted, to which Gameplay can connect and handle level transitions.

Restructured Gameplay scene to put the level scenes under the Map node so they can be contained yet we can retain an easy and reliable reference through which to manipulate them.

Added a CollisionMask value to the PhysicsShapequeryParameters2D used in detectUnits, so it won't try to select the level walls or goal.

Added refreshLevelSignals to Gameplay so that, after opening the main menu and subsequently resuming the game, the LevelCompleted signal will still be handled properly.